shallot/backend/pyproject.toml
[tool.poetry]
name = "security-onion-chat-bot"
version = "0.1.0"
description = "A chat bot interface for Security Onion"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "app", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.8.1"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
sqlalchemy = "^2.0.23"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "3.2.2"
python-multipart = "^0.0.6"
httpx = "^0.25.0"
websockets = "^12.0"
cryptography = "^41.0.0"
alembic = "^1.12.0"
python-dotenv = "^1.0.0"
aiosqlite = "^0.19.0"
email-validator = "^2.1.0"
discord-py = "^2.4.0"
slack-sdk = "^3.23.0"
botbuilder-core = "^4.14.0"  # Microsoft Teams Bot Framework
matrix-nio = {extras = ["e2e"], version = "^0.24.0"}  # Matrix client with encryption support
python-olm = "^3.2.15"  # Required for Matrix E2E encryption
python-whois = "^0.8.0"  # Required for WHOIS lookups
dnspython = "^2.4.2"  # Required for DNS lookups

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.0"
black = "^23.10.0"
flake8 = "^6.1.0"
mypy = "^1.6.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
namespace_packages = true
explicit_package_bases = true
mypy_path = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=app --cov-report=term-missing"
asyncio_mode = "auto"

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"  # See PEP 8 for why this is needed with black
